refactor: split shacl/mod.rs and trim datalog/mod.rs into sub-modules
Task 1 — shacl/mod.rs (2767 → 113 lines):
- Created src/shacl/parser.rs: SHACL Turtle parser (12 functions)
- Created src/shacl/validator.rs: Violation struct + validation engine
- Created src/shacl/af_rules.rs: sh:rule AF bridge
- Created src/shacl/spi.rs: shape persistence/loading via SPI
- Extended src/shacl/hints.rs: added DAG monitor functions
- shacl/mod.rs is now a thin re-export root (113 lines)

Task 2 — datalog/mod.rs (1685 → 665 lines):
- Replaced src/datalog/seminaive.rs stub with full semi-naive engine
  (run_inference_seminaive, run_inference_seminaive_full, run_inference,
   run_var_pred_rule, run_seminaive_inner + private helpers)
- Moved run_inference_agg to src/datalog/coordinator.rs; updated
  coordinator with real implementation (no more delegation stubs)
- Removed all moved functions from mod.rs; added pub use re-exports
  so all public symbols remain accessible at unchanged paths

All cargo check --features pg18 warnings only (no errors).
